use only one file for types
