+++ src/bmkg/types.py
@@ -1,7 +1,18 @@
-from typing import TypedDict
+from datetime import datetime
+from typing import Iterator, Literal, TypedDict
+
+from .enums import Weather
+from .schemas import (
+    Humidity,
+    Temperature,
+    WindDirection,
+    WindSpeed,
+)
 
 __all__ = [
-    "Shakemap",
+    "WeatherForecastParameter",
+    "WeatherForecastParameterId",
+    "WeatherForecastParameters",
     "EarthquakeData",
     "FeltEarthquakeData",
     "LatestEarthquakeData",
@@ -11,7 +22,32 @@ __all__ = [
     "InfoStrongEarthquakeData",
 ]
 
-Shakemap = str
+
+class WeatherForecastParameters(TypedDict, total=False):
+    datetime: Iterator[datetime]
+    hu: Iterator[Humidity]
+    humax: Iterator[Humidity]
+    humin: Iterator[Humidity]
+    t: Iterator[Temperature]
+    tmax: Iterator[Temperature]
+    tmin: Iterator[Temperature]
+    weather: Iterator[Weather]
+    wd: Iterator[WindDirection]
+    ws: Iterator[WindSpeed]
+
+
+WeatherForecastParameter = (
+    Iterator[datetime]
+    | Iterator[Humidity]
+    | Iterator[Temperature]
+    | Iterator[Weather]
+    | Iterator[WindDirection]
+    | Iterator[WindSpeed]
+)
+
+WeatherForecastParameterId = Literal[
+    "datetime", "hu", "humax", "humin", "t", "tmax", "tmin", "weather", "wd", "ws"
+]
 
 
 class EarthquakeData(TypedDict):
@@ -37,7 +73,7 @@ class FeltEarthquakeData(EarthquakeData):
 class LatestEarthquakeData(EarthquakeData):
     Potensi: str
     Dirasakan: str
-    Shakemap: Shakemap
+    Shakemap: str
 
 
 class _StrongEarthquakeData(TypedDict):
